requirements.txt
# requirements.txt
# Backend dependencies for Real-Snatcher (FastAPI, MLOps Worker)
fastapi==0.110.0
uvicorn==0.29.0
redis==5.0.4
boto3==1.34.89
torch==2.3.0
pydantic==2.7.1
aiohttp==3.9.5
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.2.1
flake8==7.0.0
black==24.4.0
